# openapi.yaml
openapi: 3.0.1
info:
  title: Ticketline
  description: This is the definition of the WebApi of Ticketline - as planned and
    implemented by Group 16 SE&PM SS2020
  contact:
    email: e11808229@student.tuwien.ac.at
  version: 0.0.1
servers:
- url: http://localhost:8080/v1
...
paths:
...
  /news:
    get:
      tags:
      - news
      summary: Return a list of news summaries
      operationId: getNewsList
      parameters:
        - in: query
          name: unreadBy
          required: false
          schema:
            type: integer
            format: int64
          description: ID of a user to list only unread news of the user
          example: 3
        - in: query
          name: page
          required: false
          schema:
            type: integer
          description: Pagination number
          example: 0
      responses:
        200:
          description: List of news summaries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NewsDTO'
        403:
          description: Not logged in
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDTO'
        404:
          description: Couldn't find user to check for unread news
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDTO'
    post:
      security:
        - ticketline_auth: []
      tags:
      - news
      summary: Create a new news entry - Admin task
      operationId: createNews
      requestBody:
        description: News entry to be persisted
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewsDTO'
        required: true
      responses:
        201:
          description: Created news entry
          content: {}
        400:
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDTO'
        403:
          description: Not logged in or insufficient permission
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDTO'
        422:
          description: Wrong credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDTO'
  /news/{newsId}:
    get:
      tags:
      - news
      summary: Get the content of one news article
      operationId: getNews
      parameters:
        - in: path
          name: newsId
          schema:
            type: integer
            format: int64
          description: ID of the news entity to return
          required: true
          example: 14
      responses:
        200:
          description: Content of the news article
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NewsDTO'
        403:
          description: Not logged in
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDTO'
        404:
          description: Couldn't find news entry with that id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDTO'
...
components:
  schemas:
...
    ErrorDTO:
      type: object
      properties:
        type:
          $ref: '#/components/schemas/ErrorType'
        message:
          type: string
          example: 'Was für ein Fehler!'
    ErrorType:
      type: string
      enum: [FATAL, WARN, INFO]
    NewsDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 14
        title:
          type: string
          example: New shiny ticketline application!
        summary:
          type: string
          example: A group of top level experts created an awesome cool app.
        content:
          type: string
          example: Bla Bla Bla Bla....
        publishedAt:
          type: string
          format: date-time
          example: 2018-06-01T07:30+08:00
        picturePath:
          type: string
          example: images/as4k_3d.png
        author:
          type: string
          example: Peter Pan
...
  securitySchemes:
    ticketline_auth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - ticketline_auth: []